Licenta/Server/requirements.txt
aiobotocore==2.15.0
botocore==1.35.16
aiohttp==3.10.5
python-dotenv==1.0.1